requirements.txt
wheel
Django
django-allauth
django-extensions
django-filter
django-modelcluster
django-taggit
django-treebeard
djangorestframework
matplotlib
numpy
opencv-python
Pillow
requests
uuid
wagtail
tensorflow